requirements.txt
rembg==1.0.0
easygui==0.98.2
Pillow==8.3.2